Actualización revision
Se crea el catalogo de Quotes
Se modifica el quotes para que lea catalogo
Se modifican campo de farm
Se modifican campos de Tecnicos
Se crea tabla para catalogo de quotes

+++ axis/axis/Models/Farm.cs
@@ -60,7 +60,7 @@ namespace AXIS.Models
         public virtual int NumberTowers { get; set; }
 
         [DisplayName("MW")]
-        public virtual int NumberMws { get; set; }
+        public virtual double NumberMws { get; set; }
 
         [DisplayName("Gearbox")]
         public virtual string Gearbox { get; set; }
